+++ src/components/interaction/loading-indicator/loading-indicator-preview.vue
@@ -1,0 +1,17 @@
+<template>
+	<preview-wrapper>
+		<template #title>
+			loading-indicator
+		</template>
+
+		<preview-section>
+			<template #title>
+				Basic indicator
+			</template>
+
+			<loading-indicator>
+				Loading data…
+			</loading-indicator>
+		</preview-section>
+	</preview-wrapper>
+</template>