+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.7.0 - 2024-08-28
 
+### New components
+
+- [`loading-indicator`](/src/components/interaction/loading-indicator/loading-indicator.md) - A simple indicator to represent loading data.
+
 ### New features
 
 - `summary-details`: A new `detailsClasses` prop has been added to allow greater control of the appearance of the details content.